+++ src/Controllers/Product/ProductAttrController.php
@@ -3,7 +3,6 @@
 namespace JtlWooCommerceConnector\Controllers\Product;
 
 use Jtl\Connector\Core\Model\Identity;
-use InvalidArgumentException;
 use Jtl\Connector\Core\Exception\TranslatableAttributeException;
 use Jtl\Connector\Core\Model\Product as ProductModel;
 use Jtl\Connector\Core\Model\TranslatableAttribute as ProductAttrModel;
@@ -11,8 +10,6 @@ use Jtl\Connector\Core\Model\TranslatableAttributeI18n as ProductAttrI18nModel;
 use JtlWooCommerceConnector\Controllers\AbstractBaseController;
 use JtlWooCommerceConnector\Utilities\Config;
 use JtlWooCommerceConnector\Utilities\SupportedPlugins;
-use WC_Product;
-use WC_Product_Attribute;
 
 class ProductAttrController extends AbstractBaseController
 {
@@ -68,6 +65,9 @@ class ProductAttrController extends AbstractBaseController
         $suppressShippingNotice = false;
         $variationPreselect     = [];
 
+        //GERMANIZED PRO
+        $food = false;
+
         /** @var  ProductAttrModel $pushedAttribute */
         foreach ($pushedAttributes as $key => $pushedAttribute) {
             foreach ($pushedAttribute->getI18ns() as $i18n) {
@@ -96,6 +96,14 @@ class ProductAttrController extends AbstractBaseController
                         }
                         if ($i18n->getName() === ProductVaSpeAttrHandlerController::GZD_MIN_AGE) {
                             $this->addOrUpdateMetaField($productId, '_min_age', $i18n->getValue());
+                        }
+                    }
+
+                    if (SupportedPlugins::isActive(SupportedPlugins::PLUGIN_WOOCOMMERCE_GERMANIZEDPRO)) {
+                        if ($i18n->getName() === ProductVaSpeAttrHandlerController::GZD_IS_FOOD) {
+                            $value = $this->util->isTrue($i18n->getValue()) ? 'yes' : 'no';
+                            $this->addOrUpdateMetaField($productId, '_is_food', $value);
+                            $food = true;
                         }
                     }
 
@@ -257,6 +265,10 @@ class ProductAttrController extends AbstractBaseController
                 \substr(ProductVaSpeAttrHandlerController::FACEBOOK_SYNC_STATUS_ATTR, 3),
                 ''
             );
+        }
+
+        if (SupportedPlugins::isActive(SupportedPlugins::PLUGIN_WOOCOMMERCE_GERMANIZEDPRO) && !$food) {
+            $this->addOrUpdateMetaField($productId, '_is_food', 'no');
         }
 
         if (!$payable) {