+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This is the changelog of our "JTL WooCommerce Connector".
 ## Unreleased
 - CO-2514 - prevent duplicate product images 
 - CO-2394 - delete removed custom attributes
+- CO-2460 - implement compatibility with germanized pro food
 
 ## 1.42.2 _2024-02-14_
 - CO-2523 - get build version from config